abunda/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai>=0.7.2
pydantic
python-dotenv